requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.22.0
plotly>=5.18.0
openpyxl>=3.1.2
matplotlib>=3.7.0
gspread
google-auth
